third_party/blink/web_tests/html/grouping_content/main-element-expected.txt

Various tests for the main element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

<main> closes <p>:
PASS main1.parentNode.nodeName == "p" is false
<p> does not close <main>:
PASS p1.parentNode.nodeName is "MAIN"
<main> can be nested inside <main>:
PASS main3.parentNode.id is "main2"
Residual style:
PASS getWeight("main4") is "700"
PASS getWeight("span1") is "700"
FormatBlock:
PASS document.getElementById("span2").parentNode.nodeName is "MAIN"
PASS successfullyParsed is true

TEST COMPLETE
